Fix: 알림 기능 수정

+++ backend/src/main/java/kr/dgucaps/capsv4/domain/notification/dto/GetNotificationStatusResponse.java
@@ -9,6 +9,6 @@ public class GetNotificationStatusResponse {
 
     private Boolean hasNotification;
 
-    private Integer notificationCount;
+    private Long notificationCount;
     
 }
